Refactored FlaskTaskr app in Blueprints - `users`, `tasks`
