+++ FlaskTaskr/flasktaskr/templates/tasks.html
@@ -3,10 +3,10 @@
 
 <h1>Welcome to FlaskTaskr</h1>
 <br>
-<a href="{{ url_for('logout') }}">Logout</a>
+<a href="{{ url_for('users.logout') }}">Logout</a>
 <div class="add-task">
   <h3>Add a new task:</h3>
-    <form action="{{ url_for('new_task') }}" method="post">
+    <form action="{{ url_for('tasks.new_task') }}" method="post">
       {{ form.csrf_token }}
       <div class="form-group">
         {{ form.name(placeholder="name") }}
@@ -66,8 +66,8 @@
           <td width="50px">{{ task.priority }}</td>
           <td width="90px">{{ task.poster.name }}</td>
           <td>
-            <a href="{{ url_for('delete_entry', task_id = task.task_id) }}">Delete</a>  -
-            <a href="{{ url_for('complete', task_id = task.task_id) }}">Mark as Complete</a>
+            <a href="{{ url_for('tasks.delete_entry', task_id = task.task_id) }}">Delete</a>  -
+            <a href="{{ url_for('tasks.complete', task_id = task.task_id) }}">Mark as Complete</a>
           </td>
         </tr>
       {% endfor %}
@@ -98,7 +98,7 @@
           <td width="50px">{{ task.priority }}</td>
           <td width="90px">{{ task.poster.name }}</td>
           <td>
-             <a href="{{ url_for('delete_entry', task_id = task.task_id) }}">Delete</a>
+             <a href="{{ url_for('tasks.delete_entry', task_id = task.task_id) }}">Delete</a>
           </td>
         </tr>
       {% endfor %}
