+++ FlaskTaskr/flasktaskr/users/forms.py
@@ -1,0 +1,27 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField
+from wtforms.validators import DataRequired, EqualTo, Length, Email
+
+
+class RegisterForm(FlaskForm):
+    name = StringField('Username', validators=[DataRequired(),
+                                               Length(min=6, max=25)])
+    email = StringField('Email', validators=[DataRequired(), Email(),
+                                             Length(min=6, max=40)])
+    password = PasswordField('Password',
+                             validators=[DataRequired(),
+                                         Length(min=6, max=40)])
+    confirm = PasswordField('Repeat Password',
+                            validators=[
+                                DataRequired(),
+                                EqualTo('password')]
+                            )
+
+
+class LoginForm(FlaskForm):
+    name = StringField(
+                       'Username',
+                       validators=[DataRequired()])
+    password = PasswordField(
+                             'Password',
+                             validators=[DataRequired()])
