+++ FlaskTaskr/tests/test_users.py
@@ -1,9 +1,9 @@
 import os
 import unittest
 
-from views import app, db
-from _config import basedir
-from models import User
+from flasktaskr import app, db
+from flasktaskr._config import basedir
+from flasktaskr.models import Task, User
 
 TEST_DB = 'test.db'
